% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_data.R
\docType{data}
\name{bank_data}
\alias{bank_data}
\title{Bank data}
\format{A data frame with x rows and y variables:
\describe{
\item{age}{age of the individual}
\item{var2}{definition 2}

}}
\usage{
bank_data
}
\description{
This is bla bla bla
}
\keyword{datasets}
