% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_message.R
\name{care_message}
\alias{care_message}
\title{Bauchi package development training}
\usage{
care_message(name)
}
\arguments{
\item{name}{The name of the message receiver.}
}
\value{
message
}
\description{
Bauchi package development training
}
\examples{
care_message("Hadley")
}
