% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infert.R
\docType{data}
\name{infert}
\alias{infert}
\title{Infertility after Spontaneous and Induced Abortion}
\format{A data frame with x rows and y variables:
\describe{
\item{var1}{definition 1}
\item{var2}{definition 2}

}}
\usage{
infert
}
\description{
This is a matched case-control study dating from before the availability of conditional logistic regression.
}
\keyword{datasets}
