% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijatables.R
\name{naijatables}
\alias{naijatables}
\title{Generate frequency tables of different variables}
\usage{
naijatables(dat, var)
}
\arguments{
\item{dat}{Dataset}

\item{var}{Variable}
}
\value{
Frequency table
}
\description{
Generate frequency tables of different variables
}
\examples{
naijatables(infert, "parity")
}
